unnecessary code removed
